# Refs #28459 -- Improved performance of Model.from_db() when fields are deferred.