Before:
```
In [3]: %timeit for x in City.objects.only('id'): pass               
1.07 s ± 10.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
After:
```
In [2]: %timeit for x in City.objects.only('id'): pass
961 ms ± 5.68 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

@timgraham, probably I'll create few more PRs to improve QuerySet performance, so I'm not sure if I should create ticket for the each one or shoud I rename [#28459](https://code.djangoproject.com/ticket/28459) to be more general and reference it.