@timgraham 2017-08-01
Feel free to reuse the ticket. The main point is to have a way to trace the commits back to their benchmarks. If you put the benchmarks and a link to the PR in a comment, that should work.